# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Add two numbers",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "operands",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.operands"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.resultResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    }
                }
            }
        },
        "/capabilities": {
            "post": {
                "description": "Returns the calculator operations supported by this service.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "List supported calculator operations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    }
                }
            }
        },
        "/divide": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Divide the first number by the second",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "operands",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.operands"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.resultResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    }
                }
            }
        },
        "/multiply": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Multiply two numbers",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "operands",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.operands"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.resultResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    }
                }
            }
        },
        "/subtract": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Subtract the second number from the first",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "operands",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.operands"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.resultResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    }
                }
            }
        },
        "/trigonometry": {
            "post": {
                "description": "Returns sin, cos, tan, cosec, sec, and cot. Undefined reciprocal values are null.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Evaluate trigonometric functions for an angle in degrees",
                "parameters": [
                    {
                        "description": "Angle in degrees",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.trigonometryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TrigonometryResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/api.errorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.TrigonometryResult": {
            "type": "object",
            "properties": {
                "cos": {
                    "type": "number"
                },
                "cosec": {
                    "type": "number"
                },
                "cot": {
                    "type": "number"
                },
                "sec": {
                    "type": "number"
                },
                "sin": {
                    "type": "number"
                },
                "tan": {
                    "type": "number"
                }
            }
        },
        "api.errorResp": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "api.operands": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "api.resultResp": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "number"
                }
            }
        },
        "api.trigonometryRequest": {
            "type": "object",
            "required": [
                "theta"
            ],
            "properties": {
                "theta": {
                    "type": "number"
                }
            }
        }
    }
}